move cdrefKeywordsExceptions.cs samples
