+++ docs/csharp/language-reference/keywords/try-catch.md
@@ -138,7 +138,7 @@ A task can be in a faulted state because multiple exceptions occurred in the awa
 
 In the following example, the `try` block contains a call to the `ProcessString` method that may cause an exception. The `catch` clause contains the exception handler that just displays a message on the screen. When the `throw` statement is called from inside `ProcessString`, the system looks for the `catch` statement and displays the message `Exception caught`.
 
-[!code-csharp[csrefKeywordsExceptions#2](~/samples/snippets/csharp/VS_Snippets_VBCSharp/csrefKeywordsExceptions/CS/csrefKeywordsExceptions.cs#2)]
+:::code language="csharp" source="./snippets/RefKeywordsExceptions.cs" id="Snippet2":::
 
 ## Two catch blocks example
 
@@ -148,7 +148,7 @@ To catch the least specific exception, you can replace the throw statement in `P
 
 If you place the least-specific catch block first in the example, the following  error message appears: `A previous catch clause already catches all exceptions of this or a super type ('System.Exception')`.
 
-[!code-csharp[csrefKeywordsExceptions#3](~/samples/snippets/csharp/VS_Snippets_VBCSharp/csrefKeywordsExceptions/CS/csrefKeywordsExceptions.cs#3)]
+:::code language="csharp" source="./snippets/RefKeywordsExceptions.cs" id="Snippet3":::
 
 ## Async method example
